+++ trunk/simLifeDisplay/src/main/java/barsuift/simLife/j3d/Axis3D.java
@@ -22,17 +22,17 @@ import javax.media.j3d.Appearance;
 import javax.media.j3d.GeometryArray;
 import javax.media.j3d.LineArray;
 import javax.media.j3d.Shape3D;
-import javax.vecmath.Point3d;
+import javax.vecmath.Point3f;
 
 public class Axis3D extends Shape3D {
 
-    private final Point3d startPoint;
+    private final Point3f startPoint;
 
-    private final Point3d endPoint;
+    private final Point3f endPoint;
 
     public Axis3D(Axis axisType) {
-        this.startPoint = new Point3d(0, 0, 0);
-        this.endPoint = new Point3d(axisType == Axis.X ? 5 : 0, axisType == Axis.Y ? 5 : 0, axisType == Axis.Z ? 5 : 0);
+        this.startPoint = new Point3f(0, 0, 0);
+        this.endPoint = new Point3f(axisType == Axis.X ? 5 : 0, axisType == Axis.Y ? 5 : 0, axisType == Axis.Z ? 5 : 0);
         createBranchShape();
     }
